requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
pydantic-settings==2.3.1
pyswisseph==2.10.3.2
geopy==2.4.1
timezonefinder==6.5.2
python-dateutil==2.9.0.post0
psycopg2-binary==2.9.9
pytest==8.2.1
httpx==0.27.0
